basic Enemy

+++ unit.js
@@ -5,6 +5,7 @@ function unit() {
     this.speed = [0, 0];
     this.auto_delete = true;
     this.disabled = false;
+    this.bounce = false;
     
     this.init = function (dom) {
         this.dom = dom;
@@ -17,7 +18,12 @@ function unit() {
         var tmpY = this.y + this.speed[1];
         //console.log(tmpX < -this.width , tmpY < -this.height , tmpX > size[0] , tmpY > size[1]);
         if (tmpX < -this.width || tmpY < -this.height || tmpX > size[0] || tmpY > size[1]) {
-            this.delete();
+            if (this.auto_delete) {
+                this.delete();
+            }
+            if (this.bounce) {
+                this.speed[0] *= -1;
+            }
         } else {
             this.dom.css("left", tmpX);
             this.dom.css("top", tmpY);
